envsetup: we need to save testeth0.nmconnection /tmp/testeth0

+++ prepare/envsetup.sh
@@ -619,7 +619,7 @@ local_setup_configure_nm_eth_part2 () {
                 yes 2>/dev/null | cp -rf /etc/sysconfig/network-scripts/ifcfg-testeth0 /tmp/testeth0
             fi
         else
-            if ! test -f /tmp/testeth0; then
+                if ! test -f /tmp/testeth0; then
                 yes 2>/dev/null | cp -rf /etc/NetworkManager/system-connections/testeth0.nmconnection /tmp/testeth0
             fi
         fi
@@ -647,9 +647,16 @@ local_setup_configure_nm_eth_part2 () {
             # THIS NEEDS TO BE DONE HERE AS DONE SEPARATELY IN VETHSETUP FOR RECREATION REASONS
             nmcli c modify testeth0 ipv4.route-metric 99 ipv6.route-metric 99
             sleep 1
-            # Copy final connection to /tmp/testeth0 for later in test usage
-            yes 2>/dev/null | cp -rf /etc/sysconfig/network-scripts/ifcfg-testeth0 /tmp/testeth0
-
+            # Copy this final profile
+            if ! test -f /tmp/nm_plugin_keyfiles; then
+                if [ ! -e /tmp/testeth0 ] ; then
+                    yes 2>/dev/null | cp -rf /etc/sysconfig/network-scripts/ifcfg-testeth0 /tmp/testeth0
+                fi
+            else
+                    if ! test -f /tmp/testeth0; then
+                    yes 2>/dev/null | cp -rf /etc/NetworkManager/system-connections/testeth0.nmconnection /tmp/testeth0
+                fi
+            fi
         fi
 
         if [ $wlan -eq 1 ]; then
